rename to plural
